vfio/core: Split IO page tbl mgmt into two files
Split IO page management into vfio and iommufd. By default we try to
build iommufd by detecting if IOMMU_IOAS_ALLOC and
VFIO_DEVICE_BIND_IOMMUFD are defined. If they are not, we build vfio
backend.

Both iommufd and vfio implementations use the vfio_container struct
which will have the corresponding elements based on the meson defined
variable HAVE_VFIO_DEVICE_BIND_IOMMUFD.

The functions in vfio.c are copied from vfio/core.c and while the ones
in iommufd.c are all new.

Moved all the logic related to io page table management related to the
iommufd infrastructure from src/vfio to src/iommufd.
