+++ src/vfio/container.h
@@ -10,22 +10,30 @@
  * COPYING and LICENSE files for more information.
  */
 
+#ifndef HAVE_VFIO_DEVICE_BIND_IOMMUFD
 struct vfio_group {
 	int fd;
 	struct vfio_container *container;
 
 	const char *path;
 };
+#endif // HAVE_VFIO_DEVICE_BIND_IOMMUFD
 
 #define VFN_MAX_VFIO_GROUPS 64
 
 struct vfio_container {
 	int fd;
-	struct vfio_group groups[VFN_MAX_VFIO_GROUPS];
 	struct iommu_state iommu;
+#ifndef HAVE_VFIO_DEVICE_BIND_IOMMUFD
+	struct vfio_group groups[VFN_MAX_VFIO_GROUPS];
+#else
+	uint32_t ioas_id;
+#endif // HAVE_VFIO_DEVICE_BIND_IOMMUFD
 };
 
 extern struct vfio_container vfio_default_container;
 
-int vfio_get_group_fd(struct vfio_container *vfio, const char *path);
 int vfio_get_device_fd(struct vfio_container *vfio, const char *bdf);
+int vfio_init_container(struct vfio_container *vfio);
+int vfio_do_map_dma(struct vfio_container *vfio, void *vaddr, size_t len, uint64_t iova);
+int vfio_do_unmap_dma(struct vfio_container *vfio, size_t len, uint64_t iova);
